Move CSStringConverter from localisation to rct12 folder (#22280)
