+++ src/openrct2/libopenrct2.vcxproj
@@ -282,7 +282,6 @@
     <ClInclude Include="interface\Window_internal.h" />
     <ClInclude Include="interface\ZoomLevel.h" />
     <ClInclude Include="Limits.h" />
-    <ClInclude Include="localisation\ConversionTables.h" />
     <ClInclude Include="localisation\Currency.h" />
     <ClInclude Include="localisation\CurrencyTypes.h" />
     <ClInclude Include="localisation\Date.h" />
@@ -388,6 +387,7 @@
     <ClInclude Include="platform\Platform.h" />
     <ClInclude Include="profiling\Profiling.h" />
     <ClInclude Include="profiling\ProfilingMacros.hpp" />
+    <ClInclude Include="rct12\CSStringConverter.h" />
     <ClInclude Include="rct12\EntryList.h" />
     <ClInclude Include="rct12\Limits.h" />
     <ClInclude Include="rct12\RCT12.h" />
@@ -811,8 +811,6 @@
     <ClCompile Include="interface\Window.cpp" />
     <ClCompile Include="interface\Window_internal.cpp" />
     <ClCompile Include="interface\ZoomLevel.cpp" />
-    <ClCompile Include="localisation\ConversionTables.cpp" />
-    <ClCompile Include="localisation\Convert.cpp" />
     <ClCompile Include="localisation\Currency.cpp" />
     <ClCompile Include="localisation\FormatCodes.cpp" />
     <ClCompile Include="localisation\Formatter.cpp" />
@@ -906,6 +904,7 @@
     <ClCompile Include="platform\Platform.Posix.cpp" />
     <ClCompile Include="platform\Platform.Win32.cpp" />
     <ClCompile Include="profiling\Profiling.cpp" />
+    <ClCompile Include="rct12\CSStringConverter.cpp" />
     <ClCompile Include="rct12\RCT12.cpp" />
     <ClCompile Include="rct12\SawyerChunk.cpp" />
     <ClCompile Include="rct12\SawyerChunkReader.cpp" />
