+++ src/openrct2/localisation/Language.h
@@ -14,7 +14,6 @@
 #include "../localisation/StringIdType.h"
 
 #include <string>
-#include <string_view>
 
 enum
 {
@@ -49,7 +48,7 @@ enum
     LANGUAGE_COUNT
 };
 
-enum class RCT2LanguageId
+enum class RCT2LanguageId : uint8_t
 {
     EnglishUK,
     EnglishUS,
@@ -89,7 +88,6 @@ uint8_t LanguageGetIDFromLocale(const char* locale);
 const char* LanguageGetString(StringId id);
 bool LanguageOpen(int32_t id);
 
-std::string RCT2StringToUTF8(std::string_view src, RCT2LanguageId languageId);
 bool LanguageGetLocalisedScenarioStrings(const utf8* scenarioFilename, StringId* outStringIds);
 void LanguageFreeObjectString(StringId stringId);
 StringId LanguageAllocateObjectString(const std::string& target);
