+++ src/openrct2/object/StringTable.cpp
@@ -13,9 +13,8 @@
 #include "../core/IStream.hpp"
 #include "../core/Json.hpp"
 #include "../core/String.hpp"
-#include "../localisation/Language.h"
-#include "../localisation/LanguagePack.h"
 #include "../localisation/LocalisationService.h"
+#include "../rct12/CSStringConverter.h"
 #include "Object.h"
 
 static constexpr uint8_t RCT2ToOpenRCT2LanguageId[] = {
