+++ src/openrct2/rct1/S4Importer.cpp
@@ -38,7 +38,6 @@
 #include "../entity/Staff.h"
 #include "../interface/Window.h"
 #include "../localisation/Date.h"
-#include "../localisation/Localisation.h"
 #include "../management/Award.h"
 #include "../management/Finance.h"
 #include "../management/Marketing.h"
@@ -49,6 +48,7 @@
 #include "../object/ObjectRepository.h"
 #include "../peep/PeepAnimationData.h"
 #include "../peep/RideUseSystem.h"
+#include "../rct12/CSStringConverter.h"
 #include "../rct12/EntryList.h"
 #include "../ride/RideData.h"
 #include "../ride/Station.h"
