+++ src/openrct2/rct12/RCT12.cpp
@@ -11,9 +11,9 @@
 
 #include "../core/String.hpp"
 #include "../localisation/Formatting.h"
-#include "../localisation/Localisation.h"
 #include "../object/ObjectList.h"
 #include "../rct1/Tables.h"
+#include "../rct12/CSStringConverter.h"
 #include "../rct2/RCT2.h"
 #include "../ride/Ride.h"
 #include "../ride/Track.h"
